+++ jupiter-rpc/src/main/java/org/jupiter/rpc/consumer/dispatcher/DefaultBroadcastDispatcher.java
@@ -24,8 +24,10 @@ import org.jupiter.rpc.consumer.future.DefaultInvokeFuture;
 import org.jupiter.rpc.consumer.future.DefaultInvokeFutureGroup;
 import org.jupiter.rpc.consumer.future.InvokeFuture;
 import org.jupiter.rpc.model.metadata.MessageWrapper;
+import org.jupiter.serialization.OutputBuf;
 import org.jupiter.serialization.Serializer;
 import org.jupiter.serialization.SerializerType;
+import org.jupiter.transport.CodecConfig;
 import org.jupiter.transport.channel.JChannel;
 import org.jupiter.transport.channel.JChannelGroup;
 
@@ -58,20 +60,31 @@ public class DefaultBroadcastDispatcher extends AbstractDispatcher {
 
         byte s_code = _serializer.code();
         // 在业务线程中序列化, 减轻IO线程负担
-        byte[] bytes = _serializer.writeObject(message);
-        request.bytes(s_code, bytes);
+        boolean isEncodeLowCopy = CodecConfig.isEncodeLowCopy();
+        if (!isEncodeLowCopy) {
+            byte[] bytes = _serializer.writeObject(message);
+            request.bytes(s_code, bytes);
+        }
 
         long invokeId = request.invokeId();
         ConsumerInterceptor[] interceptors = interceptors();
         InvokeFuture<T>[] futures = new DefaultInvokeFuture[channels.length];
         long timeoutMillis = getMethodSpecialTimeoutMillis(message.getMethodName());
         for (int i = 0; i < channels.length; i++) {
-            JChannel ch = channels[i];
+            JChannel channel = channels[i];
+
             DefaultInvokeFuture<T> future = DefaultInvokeFuture
-                    .with(invokeId, ch, returnType, timeoutMillis, DispatchType.BROADCAST)
+                    .with(invokeId, channel, returnType, timeoutMillis, DispatchType.BROADCAST)
                     .interceptors(interceptors)
                     .traceId(message.getTraceId());
-            futures[i] = write(ch, request, future, DispatchType.BROADCAST);
+
+            if (isEncodeLowCopy) {
+                OutputBuf outputBuf =
+                        _serializer.writeObject(channel.allocOutputBuf(), message);
+                request.outputBuf(s_code, outputBuf);
+            }
+
+            futures[i] = write(channel, request, future, DispatchType.BROADCAST);
         }
 
         return DefaultInvokeFutureGroup.with(futures);