+++ jupiter-serialization/jupiter-serialization-api/src/main/java/org/jupiter/serialization/Serializer.java
@@ -40,7 +40,11 @@ public abstract class Serializer {
 
     public abstract byte code();
 
+    public abstract <T> OutputBuf writeObject(OutputBuf outputBuf, T obj);
+
     public abstract <T> byte[] writeObject(T obj);
+
+    public abstract <T> T readObject(InputBuf inputBuf, Class<T> clazz);
 
     public abstract <T> T readObject(byte[] bytes, int offset, int length, Class<T> clazz);