+++ jupiter-serialization/jupiter-serialization-protostuff/src/main/java/org/jupiter/serialization/proto/ProtoStuffSerializer.java
@@ -20,14 +20,17 @@ import io.protostuff.LinkedBuffer;
 import io.protostuff.ProtostuffIOUtil;
 import io.protostuff.Schema;
 import io.protostuff.runtime.RuntimeSchema;
-import org.jupiter.common.util.Maps;
-import org.jupiter.common.util.Reflects;
+import org.jupiter.common.util.ExceptionUtil;
 import org.jupiter.common.util.SystemPropertyUtil;
 import org.jupiter.common.util.internal.InternalThreadLocal;
+import org.jupiter.serialization.InputBuf;
+import org.jupiter.serialization.OutputBuf;
 import org.jupiter.serialization.Serializer;
 import org.jupiter.serialization.SerializerType;
+import org.jupiter.serialization.proto.buffer.InputFactory;
+import org.jupiter.serialization.proto.buffer.OutputFactory;
 
-import java.util.concurrent.ConcurrentMap;
+import java.io.IOException;
 
 /**
  * Protostuff的序列化/反序列化实现, jupiter中默认的实现.
@@ -65,8 +68,6 @@ public class ProtoStuffSerializer extends Serializer {
                 .setProperty("protostuff.runtime.allow_null_array_element", allow_null_array_element);
     }
 
-    private static final ConcurrentMap<Class<?>, Schema<?>> schemaCache = Maps.newConcurrentMap();
-
     // 目的是复用 LinkedBuffer 中链表头结点 byte[]
     private static final InternalThreadLocal<LinkedBuffer> bufThreadLocal = new InternalThreadLocal<LinkedBuffer>() {
 
@@ -83,8 +84,22 @@ public class ProtoStuffSerializer extends Serializer {
 
     @SuppressWarnings("unchecked")
     @Override
+    public <T> OutputBuf writeObject(OutputBuf outputBuf, T obj) {
+        Schema<T> schema = RuntimeSchema.getSchema((Class<T>) obj.getClass());
+
+        try {
+            schema.writeTo(OutputFactory.getOutput(outputBuf), obj);
+        } catch (IOException e) {
+            ExceptionUtil.throwException(e);
+        }
+
+        return outputBuf;
+    }
+
+    @SuppressWarnings("unchecked")
+    @Override
     public <T> byte[] writeObject(T obj) {
-        Schema<T> schema = getSchema((Class<T>) obj.getClass());
+        Schema<T> schema = RuntimeSchema.getSchema((Class<T>) obj.getClass());
 
         LinkedBuffer buf = bufThreadLocal.get();
         try {
@@ -95,25 +110,29 @@ public class ProtoStuffSerializer extends Serializer {
     }
 
     @Override
-    public <T> T readObject(byte[] bytes, int offset, int length, Class<T> clazz) {
-        T msg = Reflects.newInstance(clazz, false);
-        Schema<T> schema = getSchema(clazz);
+    public <T> T readObject(InputBuf inputBuf, Class<T> clazz) {
+        Schema<T> schema = RuntimeSchema.getSchema(clazz);
+        T msg = schema.newMessage();
 
-        ProtostuffIOUtil.mergeFrom(bytes, offset, length, msg, schema);
+        try {
+            schema.mergeFrom(InputFactory.getInput(inputBuf), msg);
+        } catch (IOException e) {
+            ExceptionUtil.throwException(e);
+        } finally {
+            inputBuf.release();
+        }
+
         return msg;
     }
 
-    @SuppressWarnings("unchecked")
-    private <T> Schema<T> getSchema(Class<T> clazz) {
-        Schema<T> schema = (Schema<T>) schemaCache.get(clazz);
-        if (schema == null) {
-            Schema<T> newSchema = RuntimeSchema.createFrom(clazz);
-            schema = (Schema<T>) schemaCache.putIfAbsent(clazz, newSchema);
-            if (schema == null) {
-                schema = newSchema;
-            }
-        }
-        return schema;
+    @Override
+    public <T> T readObject(byte[] bytes, int offset, int length, Class<T> clazz) {
+        Schema<T> schema = RuntimeSchema.getSchema(clazz);
+        T msg = schema.newMessage();
+
+        ProtostuffIOUtil.mergeFrom(bytes, offset, length, msg, schema);
+
+        return msg;
     }
 
     @Override